feat: Refactor project structure and enhance functionality
- Restructure codebase from single .py file to improve maintainability
- Implement conditional logic to skip audio download and transcription when filename contains "_organized"
- Add local image storage for future upload capabilities
