+++ .gitignore
@@ -43,3 +43,5 @@ urls.txt
 # OS specific
 .DS_Store
 Thumbs.db
+temp_*/
+logs/
