+++ src/downloader/error.py
@@ -1,0 +1,11 @@
+class DownloadError(Exception):
+    """Customize download error class"""
+
+    def __init__(self, message: str, platform: str, error_type: str, details: str = None):
+        self.message = message
+        self.platform = platform
+        self.error_type = error_type
+        self.details = details
+        super().__init__(self.message)
+
+
