+++ src/logger/__init__.py
@@ -1,0 +1,3 @@
+from .logger import Logger
+
+app_logger = Logger('video-note-generator')
