+++ src/util/extend.py
@@ -1,0 +1,5 @@
+def extend(cls):
+    def decorator(func):
+        setattr(cls, func.__name__, func)
+        return func
+    return decorator
